Updates to 7
It seems as though state_ref is changing with state_new each time,
don’t think this is the problem though… h_0 is not changing, and our
RHS is increasing instead of decreasing and converging to negative h_0.
Also, I think setState and equilibrate are the ways to go, have to
think through why it’s not converging some more. don’t think its a step
size problem.